Refactor of waypoint settings into components
